chore: rename all 'workers' to 'users' for better readability
