+++ src/app/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import styles from "./styles/app.module.scss";
 import { Header } from "../components/header/header";
 import { Banner } from "../components/banner/banner";
-import { WorkersList } from "../components/workersList/workersList";
+import {UsersList} from "../components/usersList/usersList";
 
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
       <div className={styles.contentWrapper}>
         <Banner />
         <div className={styles.workersListWrapper}>
-          <WorkersList />
+          <UsersList />
         </div>
       </div>
     </div>
