+++ src/components/userCard/styles/card.module.scss
