--- src/components/workersList/styles/list.module.css
@@ -1,33 +1,0 @@
-.listContainer {
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-}
-@media screen and (max-width: 1170px) {
-  .listContainer {
-    margin: 0 60px;
-  }
-}
-@media all and (max-width: 980px) {
-  .listContainer {
-    margin: 0 32px;
-    justify-content: space-around;
-  }
-}
-@media all and (max-width: 480px) {
-  .listContainer {
-    margin: 0 16px;
-    justify-content: space-around;
-  }
-}
-
-.loadButtonContainer {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-top: 54px;
-}
-
-/*# sourceMappingURL=list.module.css.map */
